left sidebar: Fix hover title for "STREAMS".
If you click on STREAMS, it opens the
streams filter, so now the hover reflects
that.

The wording here is consistent with what
you see when you hover over USERS.

Maybe we want to say "Search ..." instead
of "Filter ..." for these two headings.

Maybe this is just a bit of a misfeature,
though, since we already have search icons,
and we don't have consistent behavior for
the GROUP PMs header.