scrollbar: Fix incorrect CSS selector.
In b6d77e3979200d3bc8743fa1d8dab9432156be41, we incorrectly changed
the selector for accessing .fixed-app .app-main to just be .app-main,
which resulted in it affecting parts of the site it should not have.

The end result was a significant layout regression of the floating
recipient bar at some widths.