docs: Rewrite translation guide to be much more user-facing.
This guide was originally written as part of the process of adding
i18n support to Zulip, so it had way too much focus on development
tooling details not relevant to translators.

There's probably a separate follow-up project we should do to move the
developers-only content to a separate article, since it's kinda lost
here for that audience.

Thanks to David Wood for asking a question that inspired me to do
this.