markdown: Make backend status message detection logic mimic frontend.
In cfc29637069349bb516101567c29012e8d15f11c, we intended to change our
markdown processor to add support for multi-line /me messages.
However, we neglected to change the backend processor, resulting in
the change only taking effect for the user sending the message :(.

We fix this by changing the backend processor too.

Fixes #12450.